feature: Remove checks based on code review
